fix whitespace

+++ dribas.sergey/T1/getShapeInfo.hpp
@@ -7,7 +7,7 @@
 
 namespace dribas
 {
-  size_t getShapeInfo(std::istream& input, std::ostream& error, Shape** myShapes, double* scalingFactor);
+  int getShapeInfo(std::istream& input, std::ostream& error, Shape** myShapes, double* scalingFactor);
   void clear(Shape ** myshape, size_t shapeCount);
   double getAllArea(const Shape* const* shapes, const size_t shapeCount);
   void scalingAll(Shape** shapes, size_t shapeCount, point_t center, double ratio);
